DatabaseSeeder: Put every table in their own files + changed how permissions will be given
